reorganized tx_update and verify_tx
