--- docs/SMPC_sidechain.md~
@@ -1,1 +1,0 @@
-SMPC will have to be a sidechain. The data is secret, so it cannot be put onto the main chain. And yet we need to maintain consensus over it in a decentralized way
