+++ docs/glossary.md
@@ -18,11 +18,13 @@ block - Used to update consensus state. Blocks contain multiple transactions. A 
 
 transaction - Explains how the consensus state should be changed. Creating transactions is how users interact with the blockchain. This is how to spend money and make bets.
 
-channel - a channel is like a small blockchain for 2 people. It allows users to use the blockchain without paying any fees.
+channel - a channel is like a small blockchain for 2 people. It allows users to use the blockchain without paying any fees. 
 
 channel state - a contract explaining who owns which parts of the money in a channel. If both participants in the channel work together, they can update to a new channel state.
 
-validator - On every block a random sample of bond holders is selected. This group of bond holders has the privelege of selecting the next block for the blockchain. They are called validators.
+mechanism - this is the thing that is being bet on. It is a true or false statment. It might be simple, like "Obama wins the 2012 election." or complicated source-code for a computer language. The computer program must return True or False.
+
+validator - On every block a random sample of bond holders is selected. This group of bond holders has the privelege of selecting the next block for the blockchain. 
 
 contract for difference - betting in flying fox uses contract for difference (CFD). It is a type of betting that happens between pairs of people. The pair of people promise to trade an amount of money based upon some future price. It is useful for hedging your risk, for example if you own 3 tons of onions and want to be protected from the possibility of onion price dropping.
