--- install.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-cd lib
-elixirc *.ex
-cd ..
-mix escript.build
