+++ lib/tx_creator.ex
@@ -1,9 +1,4 @@
 defmodule TxCreator do
-  
-  #@min_bond            Application.get_env :flying_fox, :min_bond
-  #@epoch               Application.get_env :flying_fox, :epoch
-  #@chances_per_address Application.get_env :flying_fox, :chances_per_address
-
   def nonce(pub) do
     a=Mempool.txs
     |> Enum.filter(fn(tx) -> tx.pub == pub end)
@@ -27,7 +22,7 @@ defmodule TxCreator do
         prev_hash = Blocktree.blockhash(Blockchain.get_block(h))
       end
       tot_bonds = KV.get("tot_bonds")
-      w= Enum.filter(0..Constants.chances_per_address, fn(x) -> VerifyTx.winner?(acc.bond, tot_bonds, VerifyTx.rng(prev_hash), pub, x) end) 
+      w= Enum.filter(0..Constants.chances_per_address, fn(x) -> SignTransaction.winner?(acc.bond, tot_bonds, SignTransaction.rng(prev_hash), pub, x) end) 
       h = KV.get("height") + 1
       ran = KV.get("secret #{inspect h}")
       if ran == nil do
