+++ lib/txs/win.ex
@@ -1,0 +1,7 @@
+defmodule Win do
+	def check(tx, txs) do
+		false
+	end
+	def update(tx, d) do
+	end
+end
